Implement slurm submission
